Update the validators for linting and extension
The validators have had a couple of linting changes, and the
collection validator asks for FormatterInterface.

One of the Real tests previously had a typo. This has been fixed,
and a couple of linting warnings have been fixed. Secondly, the
Collection validator's constructor now accepts any implementation
of the ValidatorInterface, rather than the ValidatorAbstract,
which should allow for more flexibility.