Add interfaces for nested validation
Previously, validation of anything would result in a one-dimensional
array of errors. By introducing the NestableResult type, we can now
create results that correlate to the validator structure. The new
`nestedValidate` method in the NestableInterface validators (which,
for now, are just `Assoc` and `Collection`) will return NestedResult
objects, which may contain further NestedResults if these validators
are used within one another. This isn't necessarily the best way to
approach this issue, but it works reasonably well for now, until
someone comes up with a better PR.