Nested validation (#9)
* Tidy class declarations and use statements

Many classes were stating extension of abstracts and implementations
of interfaces, in spite of those abstracts also implementing the
interface. This commit tidies them.

Secondly, multiple paths in the `use` declaration is a red flag
from PHPCS. This has been corrected.

* Add nested result validation

It was remarked that, for many use cases, having an object that
followed the structure of the input would be more beneficial and
easier to consume. This has now been implemented through use of
the NestableResult value object and a new nestedValidate() API.

* Add interfaces for nested validation

Previously, validation of anything would result in a one-dimensional
array of errors. By introducing the NestableResult type, we can now
create results that correlate to the validator structure. The new
`nestedValidate` method in the NestableInterface validators (which,
for now, are just `Assoc` and `Collection`) will return NestedResult
objects, which may contain further NestedResults if these validators
are used within one another. This isn't necessarily the best way to
approach this issue, but it works reasonably well for now, until
someone comes up with a better PR.